pyproject.toml
[tool.poetry]
name = "py-og-image"
version = "0.1.0"
description = "Service for generating Open Graph images from URLs"
authors = ["Antoine C <kactica.pro@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pydantic = {extras = ["email"], version = "^2.7.1"}
pydantic-settings = "^2.2.1"
selenium = "^4.20.0"
celery = {extras = ["redis"], version = "^5.4.0"}
redis = "^5.0.4" # For Celery broker/backend and potential caching
boto3 = "==1.35.27"
psycopg2-binary = "^2.9.9" # PostgreSQL driver
pillow = "^10.3.0" # For image manipulation if needed (resizing)
sqlalchemy = "^2.0.30" # ORM
alembic = "^1.13.1" # Database migrations
webdriver-manager = "^4.0.1" # To manage chromedriver
jinja2 = "^3.1.4"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.2.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 